requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
executing==1.2.0
fastjsonschema==2.18.0
flasgger==0.9.7.1
Flask==2.3.2
flask-swagger-ui==4.11.1
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
jupytext==1.15.0
kiwisolver==1.4.4
langchain==0.0.233
langsmith==0.0.5
liffpy==0.0.1a3
line-bot-sdk==3.0.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.1
mkdocs==1.5.2
mkdocs-autorefs==0.5.0
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-jupyter==0.24.2
mkdocs-material==9.1.21
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.10
Pygments==2.15.1
pyimgur==0.6.0
pymdown-extensions==10.1
PyMySQL==1.1.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.1.0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
swagger-ui-py==22.7.13
tenacity==8.2.2
tinycss2==1.2.1
toml==0.10.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
